refactor: move IDomRootChecker from application/ports to domain/ports
Fix Clean Architecture layer dependency violation.
Domain layer should not depend on application layer.
Move the port interface to domain layer so dependencies flow
correctly from outer layers (infrastructure/application) to inner (domain).
