+++ host-frontend-root/frontend-src-root/src/domain/ports/IDomRootChecker.ts
